requirements.txt
torch
torchvision
mmsegmentation
mmengine
mmcv==2.0.0rc4
openmim
opencv-python
scikit-learn
pillow
gradio
fastapi
uvicorn
python-multipart